refactor: move Logger.kt log to support
