+++ backend/src/main/kotlin/day/widdle/widdle/search/service/api/EnglishSearchApi.kt
@@ -1,6 +1,6 @@
 package day.widdle.widdle.search.service.api
 
-import day.widdle.widdle.log.logger
+import day.widdle.widdle.support.logger
 import day.widdle.widdle.search.config.ClientProperties
 import day.widdle.widdle.search.service.SearchApi
 import day.widdle.widdle.search.service.dto.WordResponse
