+++ backend/src/main/kotlin/day/widdle/widdle/word/service/WordService.kt
@@ -7,7 +7,7 @@ import day.widdle.widdle.correction.service.dto.value.CorrectionStatus.CORRECTED
 import day.widdle.widdle.event.NewWordEvent
 import day.widdle.widdle.event.publisher.WiddleEventPublisher
 import day.widdle.widdle.exception.WiddleException
-import day.widdle.widdle.log.logger
+import day.widdle.widdle.support.logger
 import day.widdle.widdle.search.service.SearchService
 import day.widdle.widdle.support.getToday
 import day.widdle.widdle.support.toJamoList
